BUG FIX: PDF file output does not complete
When writting a plot to a PDF file, gnuplot will create an empty PDF
file; e.g. a file named "test.pdf" will be created, but it will be
empty.

This appears to be an issue with running gnuplot as a subprocess with
its STDOUT being bound to a scheme input port. It's not clear why this
is an issue; i.e. gnuplot should not be using STDOUT when it writes a
PDF file.

Switching to calling gnuplot with an output pipe appears to resolve the
issue.