Changed behavior of window and pdf-file
The functions window and pdf-file now call interactive and batch,
respectively. This means that plot descriptions, such as

    (interactive (window ...))

are now simplified to

    (window ...) .

Likewise, a pdf plot is generated using the simplified expression

    (pdf-file ...) .

This also means that one should NOT explicitly call batch or
interactive. That is, a statement like

   (interactive (window ...))

will now cause an error.