Merged plot elements with "meta" elements
The language has been simplified such that one is now
effectively always using the "meta" elements. For example,
former expressions of the form

    (window (figure (cartesian (lines (<- xy)))))

are now expressed as

    ([window] ([figure] ([cartesian] ([lines] ([<-] xy)))))

where window, figure, cartesian, lines, and <- are now identified with
the former "meta" elements meta-window, meta-figure, meta-lines and
<-meta, respectively. The square brackets are just aliases to
parenthesis; it is also valid to write

    ((window) ((figure) ((cartesian) ((lines) ((<-) xy)))))

although, most will probably find this to be harder to read.

Modifying properties of plot elements is done as before. For example,
specifying line color can be done as

    ([window] ([figure] ([cartesian] ([lines color: "blue"] ([<-] xy)))))

Indenting expressions, such as

    ([window]
	 ([figure]
	  ([cartesian]
	   ([lines color: "blue"] ([<-] xy))
	   ([points color: "red"] ([<-] pq)))))

gives the appearance of a markup language.